% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-kyoto-district-data.R
\name{download_kyoto_district_data}
\alias{download_kyoto_district_data}
\title{Download the Kyoto Prefecture administrative district data}
\usage{
download_kyoto_district_data(download_dir = getwd())
}
\arguments{
\item{download_dir}{A directory where downloaded file is to saved.}
}
\value{
Path of the downloaded file (invisibly).
}
\description{
Download the Kyoto Prefecture administrative district data
provided by the Ministry of Land, Infrastructure, Transport and Tourism.
}
\examples{
\dontrun{
download_kyoto_district_data("download-dir-path")
}
}
