% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-districts.R
\name{find_districts}
\alias{find_districts}
\title{Find district by city name}
\usage{
find_districts(city_name)
}
\arguments{
\item{city_name}{Part of the city name}
}
\value{
districts of the specified city name
}
\description{
The function returns the districts of the specified city name.
The search is performed with partial matches.
}
\examples{
\dontrun{
fukuchiyama_districts <- find_districts("Fukuchiyama-shi")
}
}
