% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-fire-stations.R
\name{find_fire_stations}
\alias{find_fire_stations}
\title{Find fire stations by city name}
\usage{
find_fire_stations(city_name)
}
\arguments{
\item{city_name}{Part of the city name}
}
\value{
fire station data of the specified city name
}
\description{
The function returns the fire station data of the specified city name.
The search is performed with partial matches.
}
\examples{
\dontrun{
fukuchiyama_fire_stations <- find_fire_stations("Fukuchiyama-shi")
}
}
