% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-kyoto-districts.R
\docType{data}
\name{kyoto_districts}
\alias{kyoto_districts}
\title{Kyoto prefecture administrative district data}
\format{
Each is a tibble with 36 rows 5 variables:
\describe{
\item{city}{City names}
\item{city_code}{City codes}
\item{city_kanji}{City names in japanese}
\item{area}{Area of the city (m^2)}
\item{geom}{Geometry data (polygon or multi-polygon type) of boundaries}
}
}
\usage{
kyoto_districts
}
\description{
Information on administrative areas such as cities and wards
in Kyoto Prefecture.
}
\examples{
kyoto_districts
}
\keyword{datasets}
