% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-kyoto-fire-stations.R
\docType{data}
\name{kyoto_fire_stations}
\alias{kyoto_fire_stations}
\title{Kyoto Fire Station Data}
\format{
Each is a tibble with 122 rows 2 variables:
\describe{
\item{name}{Fire station name}
\item{geom}{Fire station point}
}
}
\usage{
kyoto_fire_stations
}
\description{
Location of fire station in Kyoto Prefecture.
The data is generated from OpenStreetMap.
}
\examples{
kyoto_fire_stations
}
\keyword{datasets}
