% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyotocities-package.R
\docType{package}
\name{kyotocities-package}
\alias{kyotocities}
\alias{kyotocities-package}
\title{kyotocities: Kyoto Prefecture administrative district data}
\description{
This package provides administrative district data of Kyoto Prefecture, Japan.
}
\author{
\strong{Maintainer}: Keisuke ANDO \email{ando@maslab.aitech.ac.jp}

}
\keyword{internal}
