% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-kyoto-district-data.R
\name{read_kyoto_district_data}
\alias{read_kyoto_district_data}
\title{Read the Kyoto Prefecture administrative district data}
\format{
Each is a tibble with 69 rows 6 variables:
\describe{
\item{city}{City names}
\item{city_code}{City codes}
\item{city_kanji}{City names in japanese}
\item{geom}{Geometry data (polygon or multi-polygon type) of boundaries}
}
}
\usage{
read_kyoto_district_data(file)
}
\arguments{
\item{file}{A file path of data to read.}
}
\value{
Information on administrative areas such as cities and wards
in Kyoto Prefecture.
}
\description{
Read the Kyoto Prefecture administrative district data.
}
\examples{
\dontrun{
kyoto_districts <- read_kyoto_district_data("N03-20230101_26_GML.zip")
}
}
